--- DevPods/Account/Account/Module/DIContainer/DIContainer.swift
@@ -1,22 +1,0 @@
-//
-//  DIContainer.swift
-//  Account
-//
-//  Created by Jeans Ruiz on 6/26/20.
-//
-
-import Foundation
-
-//final class DIContainer {
-//  
-//  private let dependencies: ModuleDependencies
-//  
-//  // Repositories here?
-//  
-//  init(dependencies: ModuleDependencies) {
-//    self.dependencies = dependencies
-//  }
-//  
-//  // Create Use Cases Here...
-//  
-//}
