+++ DevPods/Account/Account/Module/Domain/Interfaces/Repositories/AccountRepository.swift
@@ -12,8 +12,4 @@ import Shared
 public protocol AccountRepository {
   
   func getAccountDetails(session: String) -> Observable<AccountResult>
-  
-//  func markAsFavorite(session: String, userId: String, tvShowId: Int, favorite: Bool) -> Observable<StatusResult>
-//
-//  func saveToWatchList(session: String, userId: String, tvShowId: Int, watchedList: Bool) -> Observable<StatusResult>
 }
