+++ DevPods/AiringToday/AiringToday/Module/DIContainer/AiringTodayDependencies.swift
@@ -1,6 +1,6 @@
 //
-//  Module.swift
-//  PopularShows
+//  AiringTodayDependencies.swift
+//  AiringToday
 //
 //  Created by Jeans Ruiz on 6/28/20.
 //
@@ -8,7 +8,7 @@
 import Foundation
 import Networking
 
-public struct PopularShowsDependencies {
+public struct AiringTodayDependencies {
   
   let apiDataTransferService: DataTransferService
   let imagesBaseURL: String
