+++ DevPods/AiringToday/AiringToday/Module/Presentation/View/Customs/FooterReusableView.swift
@@ -10,63 +10,63 @@ import Foundation
 import UIKit
 
 class FooterReusableView: UICollectionReusableView {
-    
-    fileprivate lazy var activityIndicatorView: UIActivityIndicatorView = {
-        let activityIndicatorView = UIActivityIndicatorView()
-        activityIndicatorView.style = .whiteLarge
-        activityIndicatorView.color = .darkGray
-        activityIndicatorView.startAnimating()
-        return activityIndicatorView
-    }()
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError()
-    }
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupUI()
-    }
-    
-    // MARK: - Private
-    
-    private func setupUI() {
-        setupActivityIndicatorView()
-    }
-    
-    private func setupActivityIndicatorView() {
-        addSubview(activityIndicatorView)
-        activityIndicatorView.centerInSuperview()
-    }
-    
-    // MARK: - Public
-    
-    func startAnimating() {
-        activityIndicatorView.startAnimating()
-    }
-    
-    func stopAnimating() {
-        activityIndicatorView.stopAnimating()
-    }
+  
+  fileprivate lazy var activityIndicatorView: UIActivityIndicatorView = {
+    let activityIndicatorView = UIActivityIndicatorView()
+    activityIndicatorView.style = .whiteLarge
+    activityIndicatorView.color = .darkGray
+    activityIndicatorView.startAnimating()
+    return activityIndicatorView
+  }()
+  
+  required init?(coder aDecoder: NSCoder) {
+    fatalError()
+  }
+  
+  override init(frame: CGRect) {
+    super.init(frame: frame)
+    setupUI()
+  }
+  
+  // MARK: - Private
+  
+  private func setupUI() {
+    setupActivityIndicatorView()
+  }
+  
+  private func setupActivityIndicatorView() {
+    addSubview(activityIndicatorView)
+    activityIndicatorView.centerInSuperview()
+  }
+  
+  // MARK: - Public
+  
+  func startAnimating() {
+    activityIndicatorView.startAnimating()
+  }
+  
+  func stopAnimating() {
+    activityIndicatorView.stopAnimating()
+  }
 }
 
 extension UIView {
-    func centerInSuperview(size: CGSize = .zero) {
-        translatesAutoresizingMaskIntoConstraints = false
-        if let superviewCenterXAnchor = superview?.centerXAnchor {
-            centerXAnchor.constraint(equalTo: superviewCenterXAnchor).isActive = true
-        }
-        
-        if let superviewCenterYAnchor = superview?.centerYAnchor {
-            centerYAnchor.constraint(equalTo: superviewCenterYAnchor).isActive = true
-        }
-        
-        if size.width != 0 {
-            widthAnchor.constraint(equalToConstant: size.width).isActive = true
-        }
-        
-        if size.height != 0 {
-            heightAnchor.constraint(equalToConstant: size.height).isActive = true
-        }
+  func centerInSuperview(size: CGSize = .zero) {
+    translatesAutoresizingMaskIntoConstraints = false
+    if let superviewCenterXAnchor = superview?.centerXAnchor {
+      centerXAnchor.constraint(equalTo: superviewCenterXAnchor).isActive = true
     }
+    
+    if let superviewCenterYAnchor = superview?.centerYAnchor {
+      centerYAnchor.constraint(equalTo: superviewCenterYAnchor).isActive = true
+    }
+    
+    if size.width != 0 {
+      widthAnchor.constraint(equalToConstant: size.width).isActive = true
+    }
+    
+    if size.height != 0 {
+      heightAnchor.constraint(equalToConstant: size.height).isActive = true
+    }
+  }
 }
