+++ DevPods/KeyChainStorage/Example/KeyChainStorage/AppDelegate.swift
@@ -19,4 +19,3 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     return true
   }
 }
-
