+++ DevPods/PopularShows/PopularShows/Module/DIContainer/PopularShowsDependencies.swift
@@ -1,6 +1,6 @@
 //
-//  ModuleDependencies.swift
-//  ShowDetails
+//  PopularShowsDependencies.swift
+//  PopularShows
 //
 //  Created by Jeans Ruiz on 6/28/20.
 //
@@ -8,7 +8,7 @@
 import Foundation
 import Networking
 
-public struct ShowDetailsDependencies {
+public struct PopularShowsDependencies {
   
   let apiDataTransferService: DataTransferService
   let imagesBaseURL: String
