+++ DevPods/SearchShows/SearchShows/Module/DIContainer/SearchShowDependencies.swift
@@ -1,14 +1,15 @@
 //
-//  ModuleDependencies.swift
-//  TVShowsList
+//  SearchShowDependencies.swift
+//  TVToday
 //
-//  Created by Jeans Ruiz on 6/27/20.
+//  Created by Jeans Ruiz on 4/7/20.
+//  Copyright © 2020 Jeans. All rights reserved.
 //
 
 import Foundation
 import Networking
 
-public struct ShowListDependencies {
+public struct SearchShowDependencies {
   
   let apiDataTransferService: DataTransferService
   let imagesBaseURL: String
