+++ DevPods/Shared/Shared/Module/Data/Network/DataMapping/Genre+Decodable.swift
@@ -9,16 +9,16 @@
 import Foundation
 
 extension Genre: Decodable {
+  
+  enum CodingKeys: String, CodingKey {
+    case id
+    case name
+  }
+  
+  public init(from decoder: Decoder) throws {
+    let container = try decoder.container(keyedBy: CodingKeys.self)
     
-    enum CodingKeys: String, CodingKey {
-        case id
-        case name
-    }
-    
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-
-        self.id = try container.decode(Int.self, forKey: .id)
-        self.name = try container.decode(String.self, forKey: .name)
-    }
+    self.id = try container.decode(Int.self, forKey: .id)
+    self.name = try container.decode(String.self, forKey: .name)
+  }
 }
