+++ DevPods/Shared/Shared/Module/Data/Repositories/DefaultAccountTVShowsRepository.swift
@@ -82,4 +82,3 @@ extension DefaultAccountTVShowsRepository: AccountTVShowsRepository {
     return dataTransferService.request(endPoint, StatusResult.self)
   }
 }
-
