+++ DevPods/Shared/Shared/Module/Domain/Entities/Genre.swift
@@ -9,8 +9,8 @@
 import Foundation
 
 public struct Genre {
-    
-    public var id: Int!
-    
-    public var name: String!
+  
+  public var id: Int!
+  
+  public var name: String!
 }
