+++ DevPods/Shared/Shared/Module/Domain/UseCases/FetchShowsUseCase.swift
@@ -16,22 +16,3 @@ public protocol FetchTVShowsUseCase {
 public struct FetchTVShowsUseCaseRequestValue {
   public let page: Int
 }
-
-//// MARK: - DefaultFetchTodayShowsUseCase
-//
-//public final class DefaultFetchTVShowsUseCase: FetchTVShowsUseCase {
-//
-//  private let tvShowsRepository: TVShowsRepository
-//
-//  public init(filter: TVShowsListFilter, tvShowsRepository: TVShowsRepository) {
-//    self.filter = filter
-//    self.tvShowsRepository = tvShowsRepository
-//  }
-//
-//  public func execute(requestValue: FetchTVShowsUseCaseRequestValue) -> Observable<TVShowResult> {
-//    //let filter = requestValue.filter
-//    // TODO
-//    return tvShowsRepository.fetchShowsByGenre(genreId: 22, page: requestValue.page)
-//    //return tvShowsRepository.fetchTVShowsList(with: filter, page: requestValue.page)
-//  }
-//}
