+++ DevPods/Shared/Shared/Module/Protocols/SimpleViewState.swift
@@ -9,34 +9,34 @@
 import Foundation
 
 public enum SimpleViewState<Entity> {
-    
-    case loading
-    case paging([Entity], next: Int)
-    case populated([Entity])
-    case empty
-    case error(String)
-    
-    public var currentEntities: [Entity] {
-        switch self {
-        case .populated(let entities):
-            return entities
-        case .paging(let entities, next: _):
-            return entities
-        case .loading, .empty, .error:
-            return []
-        }
+  
+  case loading
+  case paging([Entity], next: Int)
+  case populated([Entity])
+  case empty
+  case error(String)
+  
+  public var currentEntities: [Entity] {
+    switch self {
+    case .populated(let entities):
+      return entities
+    case .paging(let entities, next: _):
+      return entities
+    case .loading, .empty, .error:
+      return []
     }
-    
-    public var currentPage: Int {
-        switch self {
-        case .loading, .populated, .empty, .error:
-            return 1
-        case .paging(_, next: let page):
-            return page
-        }
+  }
+  
+  public var currentPage: Int {
+    switch self {
+    case .loading, .populated, .empty, .error:
+      return 1
+    case .paging(_, next: let page):
+      return page
     }
-    
-    public var isInitialPage: Bool {
-        return currentPage == 1
-    }
+  }
+  
+  public var isInitialPage: Bool {
+    return currentPage == 1
+  }
 }
