+++ DevPods/TVShowsList/TVShowsList/Module/Domain/UseCases/FetchShowsByGenreTVShowsUseCase.swift
@@ -12,7 +12,7 @@ final class DefaultFetchShowsByGenreTVShowsUseCase: FetchTVShowsUseCase {
   
   private let genreId: Int
   private let tvShowsRepository: TVShowsRepository
-    
+  
   init(genreId: Int, tvShowsRepository: TVShowsRepository) {
     self.genreId = genreId
     self.tvShowsRepository = tvShowsRepository
